requirements.txt
Flask==2.3.3
anthropic==0.8.1
python-dotenv==1.0.0
mysql-connector-python==8.3.0
